Move util functions from DisplayTable/DisplayTable.js to DisplayTable/utils.js
