+++ client/src/components/DisplayTable/index.js
@@ -1,5 +1,3 @@
-export {
-  DisplayTable,
-  default_dept_name_sort_func,
-  sort_func_template,
-} from "./DisplayTable.js";
+export { DisplayTable } from "./DisplayTable.js";
+
+export { default_dept_name_sort_func, sort_func_template } from "./utils";
